pyproject.toml
[project]
name = "industry-weekly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-cloud-firestore>=2.21.0",
    "google-genai>=1.43.0",
    "langchain-google-community>=2.0.10",
    "langchain-google-genai>=2.1.12",
    "langchain-google-vertexai>=2.1.2",
    "openai>=2.3.0",
    "requests>=2.32.5",
]
